+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## v0.2.0
+## v0.3.0 (2019-09-24)
+* Moved autoload code to Listener structure
+* Created singleton Register for default Listener
+* Lazy load bindings
+
+## v0.2.0 (2019-09-13)
 * Normalised interfaces and traits
 * Added Facade plugin binding mechanism